moved Split class
